[WIP] Implement AddAccountScreen; Refactor account related files
