+++ lib/ui/views/account/new_account/new_account_view.dart
@@ -1,0 +1,1 @@
+// TODO: Dialog style
