+++ lib/ui/views/home/home_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:personal_finance_management_app/ui/views/accounts_card/accounts_card_view.dart';
 import 'package:personal_finance_management_app/ui/components/cash_flow_card.dart';
 import 'package:personal_finance_management_app/ui/components/recent_transactions_card.dart';
+import 'package:personal_finance_management_app/ui/views/account/accounts_card/accounts_card_view.dart';
 
 // View: Shows the UI to the user.
 // Single widgets also qualify as views
